+++ src/main/webapp/jqmng/scopeReconnect.js
@@ -1,33 +1,36 @@
 jqmng.define('jqmng/scopeReconnect', ['angular'], function (angular) {
 
-
     var ng = angular.module('ng');
-    ng.run(['$rootScope', function($rootScope) {
-        var _$destroy = $rootScope.$destroy;
-        $rootScope.$destroy = function() {
-            this.$$destroyed = true;
-            var res = _$destroy.apply(this, arguments);
-            this.$$nextSibling = this.$$prevSibling = null;
-        };
-        $rootScope.$reconnect = function() {
-            var child = this;
-            if (child===$rootScope) {
-                // Nothing to do here.
-                return;
-            }
-            if (!child.$$destroyed) {
-                return;
-            }
-            child.$$destroyed = false;
-            // See Scope.$new for this logic...
-            child.$$prevSibling = $rootScope.$$childTail;
-            if ($rootScope.$$childHead) {
-                $rootScope.$$childTail.$$nextSibling = child;
-                $rootScope.$$childTail = child;
-            } else {
-                $rootScope.$$childHead = $rootScope.$$childTail = child;
-            }
+    ng.config(['$provide', function($provide) {
+        $provide.decorator('$rootScope', ['$delegate', function($rootScope) {
+            var _$destroy = $rootScope.$destroy;
+            $rootScope.$destroy = function() {
+                this.$$destroyed = true;
+                var res = _$destroy.apply(this, arguments);
+                this.$$nextSibling = this.$$prevSibling = null;
+            };
+            $rootScope.$reconnect = function() {
+                var child = this;
+                if (child===$rootScope) {
+                    // Nothing to do here.
+                    return;
+                }
+                if (!child.$$destroyed) {
+                    return;
+                }
+                var parent = child.$parent;
+                child.$$destroyed = false;
+                // See Scope.$new for this logic...
+                child.$$prevSibling = parent.$$childTail;
+                if (parent.$$childHead) {
+                    parent.$$childTail.$$nextSibling = child;
+                    parent.$$childTail = child;
+                } else {
+                    parent.$$childHead = parent.$$childTail = child;
+                }
 
-        }
+            };
+            return $rootScope;
+        }]);
     }]);
 });