+++ src/main/webapp/jqmng/widgets/pageCompile.js
@@ -96,7 +96,6 @@ jqmng.define('jqmng/widgets/pageCompile', ['jquery', 'angular'], function ($, an
                 var scope = $rootScope.$new();
                 // trigger a separate page compile...
                 $compile(element)(scope);
-                // TODO check if we are not already in a $digest cycle...
                 $rootScope.$digest();
             }
             return _page.apply(this, arguments);
@@ -153,8 +152,9 @@ jqmng.define('jqmng/widgets/pageCompile', ['jquery', 'angular'], function ($, an
             }
             refreshing = true;
             // run the jquery mobile page compiler
-            // AFTER the angular compiler is completely finished.
-            // (Cannot be done in an angular directive...)
+            // AFTER the angular compiler or any linking function is completely finished.
+            // (Cannot be done in an angular directive, as this would lead to
+            // interaction problems between angular and jqm modifying the dom...)
             if (this===$rootScope) {
                 if (jqmCompilePages.length>0) {
                     var pages = jqmCompilePages;
@@ -216,10 +216,10 @@ jqmng.define('jqmng/widgets/pageCompile', ['jquery', 'angular'], function ($, an
     // that is unique for pages and dialogs.
     ng.config(['$provide', function($provide) {
         $provide.decorator('$compile', ['$delegate', function($delegate) {
+            var selector = ':jqmData(role="page"), :jqmData(role="dialog")';
+            var rolePageAttr = 'data-role-page';
             return function(element) {
-                // TODO this is not clean (also matches siblings of the current page) nor performant!
-                $('[data-role="page"]', element.parent()).attr("data-role-page", "true");
-                $('[data-role="dialog"]', element.parent()).attr("data-role-page", "true");
+                element.filter(selector).add(element.find(selector)).attr(rolePageAttr, true);
                 degradeInputs(element);
                 return $delegate.apply(this, arguments);
             }