+++ src/test/webapp/unit/eventSpec.js
@@ -1,28 +1,4 @@
 jqmng.require(['angular', "unit/testUtils"], function(angular, utils) {
-    describe("ngm:event", function() {
-
-        it('should eval the expression when the event is fired', function() {
-            var d = utils.compileInPage('<span ngm:event="{click:\'clicked=true\'}"></span>');
-            var scope = d.element.scope();
-            d.element.click();
-            expect(scope.clicked).toEqual(true);
-        });
-
-        it('should work with multiple event/function pairs', function() {
-            var d = utils.compileInPage('<span ngm:event="{mousedown:\'m = true\', click:\'c = true\'}"></span>');
-            var element = d.element;
-            var scope = element.scope();
-            expect(scope.c).toEqual(undefined);
-            element.click();
-            expect(scope.c).toEqual(true);
-            expect(scope.m).toEqual(undefined);
-            delete scope.c;
-            element.mousedown();
-            expect(scope.c).toEqual(undefined);
-            expect(scope.m).toEqual(true);
-        });
-    });
-
     describe("event shortcuts", function() {
         it("should eval the expression when the event is fired", function() {
             var events = {taphold:'taphold',swipe:'swipe', swiperight:'swiperight',