+++ src/test/webapp/unit/inputButtonSpec.js
@@ -15,7 +15,7 @@ jqmng.require(["unit/testUtils"], function(utils) {
 
         it('should use the disabled attribute', function() {
             var d = utils.compileInPage(
-                '<input type="submit" ng:click="flag = true" ng:bind-attr="{disabled: \'{{disabled}}\'}">');
+                '<input type="submit" ng:click="flag = true" disabled="{{disabled}}">');
             var input = d.element;
             var scope = input.scope();
             var parentDiv = input.parent();